pyproject.toml
[project]
name = "langchain-academy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.6.3",
    "ipywidgets>=8.1.5",
    "jupyterlab>=4.3.5",
    "langchain>=0.3.18",
    "langchain-community>=0.3.17",
    "langchain-core>=0.3.35",
    "langchain-google-genai>=2.0.9",
    "langchain-groq>=0.2.4",
    "langchain-ollama>=0.2.3",
    "langgraph>=0.2.73",
    "notebook>=7.3.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
